Capture deploy regions in code - the Dagger pipeline (#2)
We want to capture the regions that we deploy to in code, not just the
README. We also want to be explicit about how many instances we want per
region, otherwise we may end up with multiple instances in e.g. France,
and none in the US.

While at it, bump the deps that we care about - esp. latest Dagger.

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>