fix: full E2E test of our shutdown process
E2E integration test in which we test our shutdown logic to see if we lose data when having a concurrent shutdown signal and incoming streams. In this test, we setup a local HTTP and TCP server (fake syslogger), send X messages to the HTTP server and right after that send the HTTP shutdown signal (faked with a channel).

The HTTP server mimicks our main.go implementation: it wraps the handler with a waitGroup and waits for all processed logs to be written in the HTTP shutdown goroutine.

Signed-off-by: grouville <guillaume@dagger.io>